move comp tests
