+++ client/components/Home.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, mount, render } from 'enzyme';
-import Home from '../../client/components/Home';
+import Home from './Home';
 // todo: close server after tests complete
 
 describe('Home', () => {
